hack/requirements.txt
PyYAML==6.0.1
jsonschema==4.17.3